Prometheus.jl

Introduction

This package is a Julia client for Prometheus. If you are not familiar with Prometheus it is recommended to browse the upstream documentation. The documentation here focuses on the Julia client.

Two of the basic concepts of a Prometheus client are Registries and Collectors. Registries are collections of collectors, and the collectors are the units responsible to record and capture metrics. Client libraries implement a default registry which all collectors implicity register with, so for basic usage there is no need to interact with a registry (see Default registry).

The third important concept is Exposition of the collected metrics. Typically metrics are exposed over a HTTP server, as in the Quickstart-example just below. See the section about Exposition for more details and examples on how metrics can be exposed.

Quickstart

Install Prometheus.jl and HTTP.jl using the package manager:

pkg> add Prometheus HTTP

Paste the following code into a Julia REPL.

# Load the packages
using Prometheus, HTTP

# Create a Counter metric
const request_counter = Prometheus.Counter("request_count", "Number of handled requests")

# Start a HTTP server on localhost port 8000 to server the metrics
server = HTTP.listen!(8000) do http
    Prometheus.inc(request_counter) # Increment the request counter
    return Prometheus.expose(http)  # Expose the metrics
end

Visit http://localhost:8000 in your browser. You will see something like the following

# HELP gc_alloc_bytes_total Total number of allocated bytes
# TYPE gc_alloc_bytes_total counter
gc_alloc_bytes_total 365578814

[...]

# HELP request_count Number of handled requests
# TYPE request_count counter
request_count 1

The output contains some default metrics related to the running process, as well as the request counter that we added ourselves. Every time you refresh, the counter will increment its value. close(server) will shutdown the server.

Collectors

This section documents the collectors that are currently supported. This include the "basic" collectors (Counter, Gauge, Summary) as well as some custom collectors (GCCollector, ProcessCollector). There is also a section on how to implement your own collector, see Custom collectors.

Upstream documentation:

https://prometheus.io/docs/concepts/metric_types/

https://prometheus.io/docs/instrumenting/writing_clientlibs/#metrics

Counter

Quoting the upstream documentation:

A counter is a cumulative metric that represents a single monotonically increasing counter whose value can only increase or be reset to zero on restart. For example, you can use a counter to represent the number of requests served, tasks completed, or errors.

Do not use a counter to expose a value that can decrease. For example, do not use a counter for the number of currently running processes; instead use a gauge.

Counter API reference

Prometheus.Counter — Method

Prometheus.Counter(name, help; registry=DEFAULT_REGISTRY)

Construct a Counter collector.

Arguments

name :: String: the name of the counter metric.

help :: String: the documentation for the counter metric.

Keyword arguments

registry :: Prometheus.CollectorRegistry: the registry in which to register the collector. If not specified the default registry is used. Pass registry = nothing to skip registration.

Methods

Prometheus.inc: increment the counter.

Prometheus.inc — Method

Prometheus.inc(counter::Counter, v = 1)

Increment the value of the counter with v. v must be non-negative, and defaults to v = 1.

Gauge

Quoting the upstream documentation:

A gauge is a metric that represents a single numerical value that can arbitrarily go up and down.

Gauges are typically used for measured values like temperatures or current memory usage, but also "counts" that can go up and down, like the number of concurrent requests.

Gauge API reference

Prometheus.Gauge — Method

Prometheus.Gauge(name, help; registry=DEFAULT_REGISTRY)

Construct a Gauge collector.

Arguments

name :: String: the name of the gauge metric.

help :: String: the documentation for the gauge metric.

Keyword arguments

registry :: Prometheus.CollectorRegistry: the registry in which to register the collector. If not specified the default registry is used. Pass registry = nothing to skip registration.

Methods

Prometheus.inc: increment the value of the gauge.

Prometheus.dec: decrement the value of the gauge.

Prometheus.set: set the value of the gauge.

Prometheus.set_to_current_time: set the value of the gauge to the current unixtime.

Prometheus.inc — Method

Prometheus.inc(gauge::Gauge, v = 1)

Increment the value of the gauge with v. v defaults to v = 1.

Prometheus.dec — Method

Prometheus.dec(gauge::Gauge, v = 1)

Decrement the value of the gauge with v. v defaults to v = 1.

Prometheus.set — Method

Prometheus.set(gauge::Gauge, v)

Set the value of the gauge to v.

Prometheus.set_to_current_time — Method

Prometheus.set_to_current_time(gauge::Gauge)

Set the value of the gauge to the current unixtime in seconds.

Summary

Quoting the upstream documentation:

Similar to a histogram, a summary samples observations (usually things like request durations and response sizes). While it also provides a total count of observations and a sum of all observed values, it calculates configurable quantiles over a sliding time window.

Summary API reference

Prometheus.Summary — Method

Prometheus.Summary(name, help; registry=DEFAULT_REGISTRY)

Construct a Summary collector.

Arguments

name :: String: the name of the summary metric.

help :: String: the documentation for the summary metric.

Keyword arguments

registry :: Prometheus.CollectorRegistry: the registry in which to register the collector. If not specified the default registry is used. Pass registry = nothing to skip registration.

Methods

Prometheus.observe: add an observation to the summary.

Prometheus.observe — Method

Prometheus.observe(summary::Summary, v)

Add the observed value v to the summary. This increases the sum and count of the summary with v and 1, respectively.

GCCollector

A collector that exports metrics about allocations and garbage collection (for example number of allocations, number of bytes allocated, time spent in garbage collection, etc). These metrics have the julia_gc_ prefix in their name.

A GCCollector is registered automatically with the default registry, see Default registry for more details.

GCCollector API reference

Prometheus.GCCollector — Method

Prometheus.GCCollector(; registry=DEFAULT_REGISTRY)

Create a collector that exports metrics about allocations and garbage collection.

Keyword arguments

registry :: Prometheus.CollectorRegistry: the registry in which to register the collector. The default registry is used by default. Pass registry = nothing to skip registration.

Note

A GCCollector is registered automatically with the default registry. If necessary it can be removed by calling

Prometheus.unregister(Prometheus.DEFAULT_REGISTRY, Prometheus.GC_COLLECTOR)

ProcessCollector

A collector that exports metrics about a running process, for example CPU seconds and metrics about I/O operations. Metrics from this collector have the process_ prefix in their name. This collector is only available on Linux since it requires the /proc file system.

A ProcessCollector for the current process is registered automatically with the default registry, see Default registry for more details.

ProcessCollector API reference

Prometheus.ProcessCollector — Method

Prometheus.ProcessCollector(pid; registry=DEFAULT_REGISTRY)

Create a process collector for the process id given by the pid function. The collector exposes metrics about the process' CPU time, start time, memory usage, file usage, and I/O operations.

Arguments

pid :: Function: a function returning a process id as a string or integer for which to collect metrics. By default the "self" pid is used, i.e. the current process.

Keyword arguments

registry :: Prometheus.CollectorRegistry: the registry in which to register the collector. The default registry is used by default. Pass registry = nothing to skip registration.

Note

A ProcessCollector for the current process is registered automatically with the default registry. If necessary it can be removed by calling

Prometheus.unregister(Prometheus.DEFAULT_REGISTRY, Prometheus.PROCESS_COLLECTOR)

Note

The process collector is currently only available on Linux since it requires the /proc file system. On Windows and macOS this collector will not expose any metrics.

Custom collectors

RandomCollector

Labels

See https://prometheus.io/docs/concepts/data_model/#metric-names-and-labels for details.

Prometheus.Family — Method

Prometheus.Family{C}(name, help, labelnames; registry=DEFAULT_REGISTRY)

Create a labeled collector family with labels given by labelnames. For every new set of label values encountered a new collector of type C <: Collector will be created.

Arguments

name :: String: the name of the family metric.

help :: String: the documentation for the family metric.

labelnames :: Tuple{String, ...}: the label names.

Keyword arguments

registry :: Prometheus.CollectorRegistry: the registry in which to register the collector. If not specified the default registry is used. Pass registry = nothing to skip registration.

Methods

Prometheus.labels: return the collector for a specific set of labels.

Prometheus.remove: remove the collector for a specific set of labels.

Prometheus.clear: remove all collectors in the family.

Examples

# Construct a family of Counters
counter_family = Prometheus.Family{Counter}(
    "http_requests", "Number of HTTP requests", ["status_code", "endpoint"],
)

# Increment the counter for the labels status_code = "200" and endpoint = "/api"
Prometheus.inc(Prometheus.labels(counter_family, ["200", "/api"]))

Prometheus.labels — Method

Prometheus.labels(family::Family{C}, labelvalues::Tuple{String, ...}) where C

Return the collector of type C from the family corresponding to the labels given by labelvalues.

Note

This method does an acquire/release of a lock, and a dictionary lookup, to find the collector matching the label names. For typical applications this overhead does not matter (below 100ns for some basic benchmarks) but it is safe to cache the returned collector if required.

Prometheus.remove — Method

Prometheus.remove(family::Family, labelvalues::Tuple{String, ...})

Remove the collector corresponding to labelvalues. Effectively this resets the collector since Prometheus.labels will recreate the collector when called with the same label names.

Note

This method invalidates cached collectors for the label names.

Prometheus.clear — Method

Prometheus.clear(family::Family)

Remove all collectors in the family. Effectively this resets the collectors since Prometheus.labels will recreate them when needed.

Note

This method invalidates all cached collectors.

Registries

Default registry

Exposition

Prometheus support

Prometheus.expose — Function

Prometheus.expose(file::String, reg::CollectorRegistry = DEFAULT_REGISTRY)

Export all metrics in reg by writing them to the file file.

expose(io::IO, reg::CollectorRegistry = DEFAULT_REGISTRY)

Export all metrics in reg by writing them to the I/O stream io.

expose(http::HTTP.Stream, reg::CollectorRegistry = DEFAULT_REGISTRY; kwargs...)

Export all metrics in reg by writing them to the HTTP stream http.

The caller is responsible for checking e.g. the HTTP method and URI target. For HEAD requests this method do not write a body, however.